% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_ml_independent.R
\name{hb_ml_independent}
\alias{hb_ml_independent}
\title{Marginal likelihood of the non-longitudinal
independent model}
\usage{
hb_ml_independent(
  mcmc,
  data,
  response = "response",
  study = "study",
  study_reference = max(data[[study]]),
  group = "group",
  group_reference = min(data[[group]]),
  patient = "patient",
  covariates = grep("^covariate", colnames(data), value = TRUE),
  s_alpha = 30,
  s_delta = 30,
  s_beta = 30,
  s_sigma = 30,
  quiet = TRUE
)
}
\arguments{
\item{mcmc}{A wide data frame of posterior samples returned by
\code{\link[=hb_mcmc_independent]{hb_mcmc_independent()}}.}

\item{data}{A tidy data frame or \code{tibble} with the data.}

\item{response}{Character of length 1,
name of the column in \code{data} with the response/outcome variable.
\code{data[[response]]} must be a continuous variable,
and it \emph{should} be the change from baseline of a
clinical endpoint of interest, as opposed to just
the raw response. Treatment differences
are computed directly from this scale, please supply
change from baseline unless you are absolutely certain
that treatment differences computed directly from
this quantity are clinically meaningful.}

\item{study}{Character of length 1,
name of the column in \code{data} with the study ID.}

\item{study_reference}{Atomic of length 1,
element of the \code{study} column that indicates
the current study.
(The other studies are historical studies.)}

\item{group}{Character of length 1,
name of the column in \code{data} with the group ID.}

\item{group_reference}{Atomic of length 1,
element of the \code{group} column that indicates
the control group.
(The other groups may be treatment groups.)}

\item{patient}{Character of length 1,
name of the column in \code{data} with the patient ID.}

\item{covariates}{Character vector of column names
in \code{data} with the columns with baseline covariates.
These can be continuous, categorical, or binary.
Regardless, \code{historicalborrow} derives the appropriate
model matrix.}

\item{s_alpha}{Numeric of length 1, prior standard deviation
of the study-specific control group mean parameters \code{alpha}.}

\item{s_delta}{Numeric of length 1, prior standard deviation
of the study-by-group effect parameters \code{delta}.}

\item{s_beta}{Numeric of length 1, prior standard deviation
of the fixed effects \code{beta}.}

\item{s_sigma}{Numeric of length 1, prior upper bound
of the residual standard deviations.}

\item{quiet}{Logical of length 1, \code{TRUE} to suppress R console output.}
}
\value{
A numeric of length 1 with the marignal likelihood from bridge
sampling, with an object of class \code{"bridge"}
from the \code{bridgesampling}
package assigned as an attribute called \code{"bridge"}.
}
\description{
Estimate the marginal likelihood of the
non-longitudinal independent model using bridge sampling
(\code{bridgesampling} R package).
}
\details{
Bridge sampling is described by Gronau et al. (2020).
}
\examples{
if (!identical(Sys.getenv("HB_TEST", unset = ""), "")) {
data <- hb_sim_independent(n_continuous = 2)$data
mcmc <- hb_mcmc_independent(
  data,
  n_chains = 1,
  n_adapt = 100,
  n_warmup = 50,
  n_iterations = 50 # need a lot more in real life
)
suppressWarnings(hb_ml_independent(mcmc = mcmc, data = data))
}
}
\references{
Gronau QF, Singmann H, Wagenmakers E (2020). “bridgesampling:
An R Package for Estimating Normalizing Constants.” \emph{Journal of
Statistical Software}, \emph{92}(10), 1-29.
doi:10.18637/jss.v092.i10
\url{https://doi.org/10.18637/jss.v092.i10}.
}
\seealso{
Other marginal-likelihood: 
\code{\link{hb_ml_pool}()}
}
\concept{marginal-likelihood}
