% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_ml_pool.R
\name{hb_ml_pool}
\alias{hb_ml_pool}
\title{Marginal likelihood of the non-longitudinal pooled model}
\usage{
hb_ml_pool(
  mcmc,
  data,
  response = "response",
  study = "study",
  study_reference = max(data[[study]]),
  group = "group",
  group_reference = min(data[[group]]),
  patient = "patient",
  covariates = grep("^covariate", colnames(data), value = TRUE),
  s_alpha = 30,
  s_delta = 30,
  s_beta = 30,
  s_sigma = 30,
  quiet = TRUE
)
}
\arguments{
\item{mcmc}{A wide data frame of posterior samples returned by
\code{\link[=hb_mcmc_pool]{hb_mcmc_pool()}}.}

\item{data}{A tidy data frame or \code{tibble} with the data.}

\item{response}{Character of length 1,
name of the column in \code{data} with the response/outcome variable.
\code{data[[response]]} must be a continuous variable,
and it \emph{should} be the change from baseline of a
clinical endpoint of interest, as opposed to just
the raw response. Treatment differences
are computed directly from this scale, please supply
change from baseline unless you are absolutely certain
that treatment differences computed directly from
this quantity are clinically meaningful.}

\item{study}{Character of length 1,
name of the column in \code{data} with the study ID.}

\item{study_reference}{Atomic of length 1,
element of the \code{study} column that indicates
the current study.
(The other studies are historical studies.)}

\item{group}{Character of length 1,
name of the column in \code{data} with the group ID.}

\item{group_reference}{Atomic of length 1,
element of the \code{group} column that indicates
the control group.
(The other groups may be treatment groups.)}

\item{patient}{Character of length 1,
name of the column in \code{data} with the patient ID.}

\item{covariates}{Character vector of column names
in \code{data} with the columns with baseline covariates.
These can be continuous, categorical, or binary.
Regardless, \code{historicalborrow} derives the appropriate
model matrix.}

\item{s_alpha}{Numeric of length 1, prior standard deviation
of the study-specific control group mean parameters \code{alpha}.}

\item{s_delta}{Numeric of length 1, prior standard deviation
of the study-by-group effect parameters \code{delta}.}

\item{s_beta}{Numeric of length 1, prior standard deviation
of the fixed effects \code{beta}.}

\item{s_sigma}{Numeric of length 1, prior upper bound
of the residual standard deviations.}

\item{quiet}{Logical of length 1, \code{TRUE} to suppress R console output.}
}
\value{
An object of class \code{"bridge"} from the \code{bridgesampling}
package. The \code{logml} element has the actual
estimated marginal log likelihood.
}
\description{
Estimate the marginal likelihood of the
non-longitudinal pooled model using bridge sampling
(\code{bridgesampling} R package).
}
\examples{
if (!identical(Sys.getenv("HB_TEST", unset = ""), "")) {
data <- hb_sim_pool(n_continuous = 2)$data
mcmc <- hb_mcmc_pool(
  data,
  n_chains = 1,
  n_adapt = 100,
  n_warmup = 50,
  n_iterations = 50 # need a lot more in real life
)
suppressWarnings(hb_ml_pool(mcmc = mcmc, data = data))
}
}
\seealso{
Other marginal-likelihood: 
\code{\link{hb_ml_independent}()}
}
\concept{marginal-likelihood}
