% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_plot_weighted.R
\name{hb_plot_weighted}
\alias{hb_plot_weighted}
\title{Plot weighted averages between the independent and pooled models.}
\usage{
hb_plot_weighted(
  summary,
  outcome = c("response", "diff"),
  highlight_weights = numeric(0)
)
}
\arguments{
\item{summary}{Data frame generated by \code{\link[=hb_summary_weighted]{hb_summary_weighted()}}.}

\item{outcome}{Character of length 1, either \code{"response"}
or \code{"diff"}, the quantity to plot on the vertical axis.}

\item{highlight_weights}{Numeric vector of values between 0 and 1,
weights to highlight in bold blue in the plot.}
}
\value{
A \code{ggplot} object
}
\description{
Plot weighted averages between
the independent and pooled models.
}
\examples{
if (!identical(Sys.getenv("HB_TEST", unset = ""), "")) {
set.seed(1)
data <- hb_sim_independent(n_continuous = 2)$data
data$group <- paste0("group_", data$group)
mcmc_pool <- hb_mcmc_pool(
  data,
  n_chains = 1,
  n_adapt = 100,
  n_warmup = 200,
  n_iterations = 200
)
mcmc_independent <- hb_mcmc_independent(
  data,
  n_chains = 1,
  n_adapt = 100,
  n_warmup = 200,
  n_iterations = 200
)
summary <- hb_summary_weighted(
  mcmc_pool = mcmc_pool,
  mcmc_independent = mcmc_independent,
  data = data,
  weights = c(0, 0.25, 0.5, 0.75, 1)
)
hb_plot_weighted(
  summary = summary,
  outcome = "response",
  highlight_weights = 0.5
)
}
}
\seealso{
Other plot: 
\code{\link{hb_plot_borrow}()},
\code{\link{hb_plot_group}()},
\code{\link{hb_plot_tau}()}
}
\concept{plot}
