% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_s_tau.R
\name{hb_s_tau}
\alias{hb_s_tau}
\title{Suggest s_tau}
\usage{
hb_s_tau(precision_ratio = 0.5, sigma = 1, n = 100)
}
\arguments{
\item{precision_ratio}{Positive numeric vector of elements between 0 and 1
with target precision ratios.}

\item{sigma}{Positive numeric vector of residual standard deviations.}

\item{n}{Number of non-missing patients.}
}
\value{
Numeric of length equal to \code{length(precision_ratio)} and
\code{length(sigma)}, suggested values of s_tau for each element of
\code{precision_ratio} and \code{sigma}.
}
\description{
Suggest a value of the \code{s_tau} hyperparameter
to achieve a given amount of borrowing in the hierarchical model.
}
\details{
See the hierarchical model section of the
methods vignette for details.
}
\examples{
hb_s_tau(precision_ratio = 0.5, sigma = 1, n = 100)
}
\seealso{
Other data: 
\code{\link{hb_data}()}
}
\concept{data}
