% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_sim_hierarchical.R
\name{hb_sim_hierarchical}
\alias{hb_sim_hierarchical}
\title{Non-longitudinal hierarchical simulations.}
\usage{
hb_sim_hierarchical(
  n_study = 5,
  n_group = 3,
  n_patient = 100,
  n_continuous = 0,
  n_binary = 0,
  s_delta = 1,
  s_beta = 1,
  s_sigma = 1,
  s_mu = 1,
  s_tau = 1,
  alpha = stats::rnorm(n = n_study, mean = mu, sd = tau),
  delta = stats::rnorm(n = n_group - 1, mean = 0, sd = s_delta),
  beta = stats::rnorm(n = n_study * (n_continuous + n_binary), mean = 0, sd = s_delta),
  sigma = stats::runif(n = n_study, min = 0, max = s_sigma),
  mu = stats::rnorm(n = 1, mean = 0, sd = s_mu),
  tau = stats::runif(n = 1, min = 0, max = s_tau)
)
}
\arguments{
\item{n_study}{Number of studies to simulate.}

\item{n_group}{Number of groups (e.g. study arms) to simulate per study.}

\item{n_patient}{Number of patients to simulate per study per group.}

\item{n_continuous}{Number of continuous covariates to simulate
(all from independent standard normal distributions).}

\item{n_binary}{Number of binary covariates to simulate
(all from independent Bernoulli distributions with p = 0.5).}

\item{s_delta}{Numeric of length 1, prior standard deviation
of the study-by-group effect parameters \code{delta}.}

\item{s_beta}{Numeric of length 1, prior standard deviation
of the fixed effects \code{beta}.}

\item{s_sigma}{Numeric of length 1, prior upper bound
of the residual standard deviations.}

\item{s_mu}{Numeric of length 1,
prior standard deviation of \code{mu}.}

\item{s_tau}{Numeric of length 1,
Upper bound on \code{tau}.}

\item{alpha}{Numeric vector of length 1 for the pooled
and mixture models and length \code{n_study} for the
independent and hierarchical models.
\code{alpha} is the vector of control group mean parameters.
\code{alpha} enters the model by multiplying with
\verb{$matrices$x_alpha} (see the return value).
The control group in the data is the one with the
\code{group} column equal to 1.}

\item{delta}{Numeric vector of length \code{n_group - 1}
of treatment effect parameters.
\code{delta} enters the model by multiplying with
\verb{$matrices$x_delta} (see the return value).
The control (non-treatment) group in the data is the one with the
\code{group} column equal to 1.}

\item{beta}{Numeric vector of \code{n_study * (n_continuous + n_binary)}
fixed effect parameters. Within each study,
the first \code{n_continuous} betas
are for the continuous covariates, and the rest are for
the binary covariates. All the \code{beta}s for one study
appear before all the \code{beta}s for the next study,
and studies are arranged in increasing order of
the sorted unique values in \verb{$data$study} in the output.
\code{betas} enters the model by multiplying with
\verb{$matrices$x_alpha} (see the return value).}

\item{sigma}{Numeric vector of \code{n_study} study-specific
residual standard deviations.}

\item{mu}{Numeric of length 1,
mean of the control group means \code{alpha}.}

\item{tau}{Numeric of length 1,
standard deviation of the control group means \code{alpha}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{data}: tidy long-form dataset with the patient-level data.
one row per patient and indicator columns for the study,
group (e.g. treatment arm), and patient ID. The \code{response}
columns is the patient response. The other columns are
baseline covariates. The control group is the one with
the \code{group} column equal to 1, and the current study (non-historical)
is the one with the maximum value of the \code{study} column.
Only the current study has any non-control-group patients,
the historical studies have only the control group.
\item \code{parameters}: named list of model parameter values.
See the model specification vignette for details.
\item \code{matrices}: A named list of model matrices.
See the model specification vignette for details.
}
}
\description{
Simulate from the non-longitudinal hierarchical model.
}
\examples{
hb_sim_hierarchical()$data
}
\seealso{
Other simulate: 
\code{\link{hb_sim_independent}()},
\code{\link{hb_sim_mixture}()},
\code{\link{hb_sim_pool}()}
}
\concept{simulate}
