% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_package.R
\docType{package}
\name{historicalborrow-package}
\alias{historicalborrow-package}
\title{historicalborrow: Bayesian historical borrowing models for clinical studies.}
\description{
Bayesian historical borrowing models for clinical studies.
}
\concept{help}
