% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_weight_bma.R
\name{hb_weight_bma}
\alias{hb_weight_bma}
\title{Bayesian model averaging weight for the pooled model
using the method by Best et al. (2019).}
\usage{
hb_weight_bma(mll_pool, mll_independent, prior_weight = 0.5)
}
\arguments{
\item{mll_pool}{Finite numeric of length 1,
marginal log likelihood of the pooled model.
Compute with \code{\link[=hb_mll_pool]{hb_mll_pool()}}.}

\item{mll_independent}{Finite numeric of length 1,
marginal log likelihood of the pooled model.
Compute with \code{\link[=hb_mll_independent]{hb_mll_independent()}}.}

\item{prior_weight}{Numeric of length 1 between 0 and 1,
prior weight on the pooled model.}
}
\description{
Calculate the weight on the pooled model
using the Bayesian model averaging technique by
Best et al (2019).
}
\details{
This calcualtion is part of a Bayesian model averaging
placebo borrowing method which averages the posteriors
of the pooled and independnet models.
Uses the method by Best et al. (2019). See the vignettes
for details.
}
\examples{
if (!identical(Sys.getenv("HB_TEST", unset = ""), "")) {
data <- hb_sim_independent(n_continuous = 2)$data
mcmc_pool <- hb_mcmc_pool(
  data,
  n_chains = 1,
  n_adapt = 2000,
  n_warmup = 2000,
  n_iterations = 8000
)
mcmc_independent <- hb_mcmc_independent(
  data,
  n_chains = 1,
  n_adapt = 2000,
  n_warmup = 2000,
  n_iterations = 8000
)
suppressWarnings({
  mll_pool <- hb_mll_pool(mcmc = mcmc_pool, data = data)
  mll_independent <- hb_mll_independent(mcmc = mcmc_independent, data = data)
})
hb_weight_bma(
  mll_pool = mll_pool,
  mll_independent = mll_independent
)
}
}
\references{
Best N, Price RG, Pouliquen IJ, Keene ON.
Assessing efficacy in important subgroups in confirmatory trials:
An example using Bayesian dynamic borrowing.
Pharm Stat. 2021 May;20(3):551-562. doi: 10.1002/pst.2093.
Epub 2021 Jan 21. PMID: 33475231; PMCID: PMC8247867.
\url{https://onlinelibrary.wiley.com/doi/10.1002/pst.2093}.
}
\concept{weights}
